Frontend: split the code into modules, change function on arrow fucntion, add destructurization in function
