+++ index.html
@@ -95,6 +95,6 @@
       <div class="wall-left"></div>
       <div class="wall-right"></div>
     </div>
-    <script src="main.js"></script>
+    <script src="main.js" type="module"></script>
   </body>
 </html>
